To fix errors associated with using the surfe library in another application - related to exporting STL data members for exported classes.
TO fix the errors I initialized these data members in the constructors as pointers. Updated the rest of the code to be compliant with this.

Also fixed other warnings that the compiler was giving me. There are still a lot more warnings related to the gmp/mpir library and templated names exceeding 4096 chars. Looking online it seems that these warnings are okay to ignore. I always got these warnings in the gocad plugin and I had no problems.