Changed MPIR to GMP library
Note MPIR is the GMP library but customized for windows platforms. GMP is better to use since it can be used on any platform. In windows it is harder to build the library but it can be done. I have built this library on windows and included the bin/include/lib files x64 only. If you have a 32 bit machine - go set it on fire, and get a 64 bit machine.