Performed a lot of testing here. Issues: Oncontact uncertainty, supplied as a real 3D distance, might not be honored. e.g. scalar field s(x) might not be consistent with distance - the problem is a consequence of the ||Vs(x)|| != 1 - one  solution could be to try and supply the constraint ||Vs(x)|| == 1. But this has its difficulties (nonlinear and I don't know how to add it to the QP system). However, in practice you really don't want to do this because nature doesn't behave this way (hinge thickening and limb thinning). - how to ensure oncontact uncertainty?  -> after qp solution is found measure scalar field @ oncontact pts.  -> for points that violate oncontact uncertainty ( |s(x)| > oncontact uncertainty d ) modify bounds - how do you measure d (distance b/t point and nearest pt on surface)?  -> measure the negative of the gradient of that point and step in that direction until you reach surface; measure path.
Tried using the inner product (instead of the normal bound constraints for the gradient constraints) to see results. Changing the angle bounds do not really change the results. Very nice smooth results though. This is because the norm of the gradient is wrapped in the rhs (angle bound constraint) and is not being taking into consideration. Tried updating to account for the norm of the gradient in every iteration of the qpp. Convergence an issue.