Lots of updates here: More development of the restricted range functionality New Feature (should have made a few branch): conversion of the modified kernel solution to the normal rbf kernel solution (massive performance benefits) - Still needs a lot more testing....
Other notes: wasn't really happy with the results with converting normal constraints to tangent constraints in the restricted range approach via the inner product constraint. changing values of the inner product constraint did not really affect the solution too much. This problem is related to us not knowing the value of the gradient norm prior to solving system.